Update groundwater_capn2016.R
Change the derivative functions to be (forward) finite difference approximations. This is much simpler than trying to calculate them analytically. The profit and water withdrawal are the ones that were done wrong, but I've included for the Crop choice function as well for consistency. I also changed a line to pull the data file from my clone repository, not Fenichel's original.