Update groundwater_capn2016.R
I update the profit function to account for reduced profits that must follow from reduced yields in the face of deficit irrigation. This effect is entirely absent in Fenichel et al 2016. I do this in a very ad hoc way (but it must be done, ignoring it entirely doesn't make the biology of evapotranspiration magically disappear).  I leave profits when water stock and use are above 1.75 feet alone. I say profits at zero are going to be half of what they are at 1.75 (this is not a bad approximation for the profit per acre hit between irrigated and unirrigated row crops). Then I set that the profit at any point in between 0 and 1.75 feet to be the linear interpolation of the values at 0 and 1.75. The way this is done is not very important. The farmer will only have one season (approximately, this is a Cont-time model) in this range of water levels (then they are switched to nonirrigated) The most important aspect is that there is a realistic value for the non-irrigated values that are assumed to continue forever. This drop in profits is the main theoretical underpinning for a shadow price of water. Marginal water can allow for one more season of irrigation. The more water stock you have the less valuable is marginal water because this additional season of irrigated farming happens very far in the future.